little improve to login modal; added button that sends that modal
