+++ bot/ui/__init__.py
@@ -1,1 +1,2 @@
-from .discordModal import LoginModal
+from .loginModalComponent import LoginModalComponent
+from .loginButtonComponent import LoginButtonComponent
